feat(archiver): base structure for pi_worker
