+++ mlh_archiver/Cargo.toml
@@ -28,6 +28,7 @@ nntp = { path = "./rust-nntp" }
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.9"
 thiserror = "2.0"
+gix = "0.81.0"
 
 [dev-dependencies]
 testcontainers = { version =  "0.27" , features = ["blocking"] }
